build: resolve merge issue conflicts

+++ src/utils/wallet.ts
@@ -117,7 +117,10 @@ const deleteWalletWithAccounts = async (hash: string) => {
   }
 };
 
-export const importWalletViaPrivateKey = async (privateKey: string) => {
+export const importWalletViaPrivateKey = async (
+  privateKey: string,
+  accounts: AccountDBModel[]
+) => {
   let walletInDb: WalletDBModel | null = null;
   let accountInDb: AccountDBModel | null = null;
   const currencyCode = CryptoCurrencyCode.AMB;
@@ -141,8 +144,11 @@ export const importWalletViaPrivateKey = async (privateKey: string) => {
       privateKey
     );
 
-    walletInDb = await WalletDB.createWallet(fullWallet);
+    if (!_account) throw new Error();
 
+    await API.watcherService.watchAddresses([_account.address]);
+
+    walletInDb = await WalletDB.createWallet(fullWallet);
     const [, accountInDbResult] = await Promise.all([
       // Securely store private key
       Cache.setItem(
@@ -160,13 +166,19 @@ export const importWalletViaPrivateKey = async (privateKey: string) => {
     ]);
 
     accountInDb = accountInDbResult;
-    // subscribe to notifications
+
     await API.watcherService.watchAddresses([_account.address]);
-    return { hash };
+    return { hash, address: _account.address };
   } catch (error) {
-    if (walletInDb) walletInDb.destroyPermanently();
-    if (accountInDb) accountInDb.destroyPermanently();
-    throw error;
+    const _account = await AirDAOKeysForRef.discoverAccountViaPrivateKey(
+      privateKey
+    );
+
+    if (AccountUtils.isWalletAreadyExist(_account.address, accounts)) {
+      if (walletInDb) walletInDb.destroyPermanently();
+      if (accountInDb) accountInDb.destroyPermanently();
+      throw error;
+    }
   }
 };
